update and delete image

+++ app/Http/Controllers/PostController.php
@@ -42,7 +42,7 @@ class PostController extends Controller implements HasMiddleware
     public function store(Request $request)
     {
         // Validate the request
-        $fields = $request->validate([
+        $request->validate([
             'title' => ['required', 'max:255'],
             'body' => ['required'],
             'image' => ['nullable', 'file', 'max:3000', 'mimes:jpeg,png,jpg']
@@ -86,13 +86,26 @@ class PostController extends Controller implements HasMiddleware
     public function update(Request $request, Post $post)
     {
         // Validate the request
-        $fields = $request->validate([
+        $request->validate([
             'title' => ['required', 'max:255'],
             'body' => ['required'],
+            'image' => ['nullable', 'file', 'max:3000', 'mimes:jpeg,png,jpg'],
         ]);
 
+        $path = $post->image ?? null;
+        if ($request->hasFile('image')) {
+            if ($post->image) {
+                Storage::disk('public')->delete($post->image);
+            }
+            $path = Storage::disk('public')->put('posts_images', $request->image);
+        }
+
         // Update the post
-        $post->update($fields);
+        $post->update([
+            'title'=> $request->title,
+            'body'=> $request->body,
+            'image'=> $path
+        ]);
         return redirect()->route('dashboard')->with('success','Your post has been updated successfully.');
     }
 
@@ -101,6 +114,10 @@ class PostController extends Controller implements HasMiddleware
      */
     public function destroy(Post $post)
     {
+        Gate::authorize('modify', $post); 
+        if ($post->image) {
+            Storage::disk('public')->delete($post->image);
+        }
         $post->delete();
         return back()->with('delete', 'Your post has been deleted successfully.');
     }
